+++ src/components/data-processing/MitochondrialContent/MitochondrialContent.jsx
@@ -1,7 +1,6 @@
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
-
-import React from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import _ from 'lodash';
 import {
   Collapse, Row, Col, Space, Tooltip, Button,
 } from 'antd';
@@ -9,531 +8,168 @@ import PropTypes from 'prop-types';
 import {
   InfoCircleOutlined,
 } from '@ant-design/icons';
-import _ from 'lodash';
-import { Vega } from 'react-vega';
+import Loader from '../../Loader';
 import plot1Pic from '../../../../static/media/plot3.png';
 import plot2Pic from '../../../../static/media/plot4.png';
-import plotData from './data2.json';
 
+import {
+  updatePlotConfig,
+  loadPlotConfig,
+  savePlotConfig,
+} from '../../../redux/actions/componentConfig';
+
+import MitochondrialFractionHistogram from '../../plots/mitochondrialFractionHistogram';
+import MitochondrialFractionLogHistogram from '../../plots/mitochondrialFractionLogHistogram';
+
+import PlotStyling from '../../plots/styling/PlotStyling';
 import CalculationConfig from './CalculationConfig';
 
-import OldPlotStyling from '../../plots/styling/OldPlotStyling';
-
 const { Panel } = Collapse;
-class MitochondrialContent extends React.Component {
-  constructor(props) {
-    super(props);
-    this.defaultConfig = {
-      plotToDraw: true,
-      data: plotData,
-      legendEnabled: true,
-      xAxisText: 'Fraction of mitochondrial reads',
-      yAxisText: 'Fraction of cells',
-      xAxisText2: 'log10(#UMIs in cell)',
-      yAxisText2: 'Fraction of mitochondrial reads',
-      xDefaultTitle: 'Fraction of mitochondrial reads',
-      yDefaultTitle: 'Fraction of cells',
-      legendPosition: 'top-right',
-      gridWeight: 0,
-      titleSize: 12,
-      titleText: '',
-      titleAnchor: 'start',
-      masterFont: 'sans-serif',
-      masterSize: 13,
-      maxFraction: 0.1,
-      maxFraction2: 3.5,
-      axisTitlesize: 13,
-      axisTicks: 13,
-      axisOffset: 0,
-      transGrid: 0,
-      width: 530,
-      height: 400,
-      maxWidth: 660,
-      maxHeight: 560,
-      placeholder: 0.1,
-      sliderMax: 1,
-      sliderMin: 0,
-      binStep: 0.05,
-    };
-    this.state = {
-      config: _.cloneDeep(this.defaultConfig),
-      data: plotData,
-    };
-    this.updatePlotWithChanges = this.updatePlotWithChanges.bind(this);
-  }
+const MitochondrialContent = (props) => {
+  const {
+    experimentId,
+  } = props;
 
-  generateData() {
-    const { data } = this.state;
-    return data;
-  }
+  const dispatch = useDispatch();
 
-  updatePlotWithChanges(obj) {
-    this.setState((prevState) => {
-      const newState = _.cloneDeep(prevState);
+  const [selectedPlot, setSelectedPlot] = useState('histogram');
+  const [plot, setPlot] = useState(false);
 
-      _.merge(newState.config, obj);
+  const debounceSave = useCallback(_.debounce((plotUuid) => dispatch(savePlotConfig(experimentId, plotUuid)), 2000), []);
 
-      return newState;
-    });
-  }
+  const updatePlotWithChanges = (obj) => {
+    dispatch(updatePlotConfig(plots[selectedPlot].plotUuid, obj));
+    debounceSave(plots[selectedPlot].plotUuid);
+  };
 
-  generateSpec() {
-    const { config } = this.state;
-    let legend = null;
-    const colorExpression = `(datum.bin1 <= ${config.maxFraction}) ? 'Alive' : 'Dead'`;
-    const colorExpression2 = '(datum.bin1 <= 2.5) ? \'Dead\' : (datum.bin1 >=3.5) ? \'Live\' : \'Unknown\'';
-    if (config.legendEnabled) {
-      legend = [
-        {
-          fill: 'color',
-          orient: config.legendPosition,
-          labelFont: config.masterFont,
-          titleFont: config.masterFont,
-          encode: {
-            title: {
-              update: {
-                fontSize: { value: 14 },
-              },
-            },
-            labels: {
-              interactive: true,
-              update: {
-                fontSize: { value: 12 },
-                fill: { value: 'black' },
-              },
-            },
-            symbols: {
-              update: {
-                stroke: { value: 'transparent' },
-              },
-            },
-            legend: {
-              update: {
-                stroke: { value: '#ccc' },
-                strokeWidth: { value: 1.5 },
-              },
-            },
-          },
-        },
-      ];
-    } else {
-      legend = null;
+  const plots = {
+    histogram: {
+      title: 'Mitochondrial Fraction',
+      imgSrc: plot1Pic,
+      plotUuid: 'mitochondrialFractionHistogram',
+      plotType: 'mitochondrialFractionHistogram',
+      plot: (config, plotData) => (<MitochondrialFractionHistogram experimentId={experimentId} config={config} plotData={plotData} />),
+    },
+    logHistogram: {
+      title: 'Mitochondrial Fraction (Log)',
+      imgSrc: plot2Pic,
+      plotUuid: 'mitochondrialFractionLogHistogram',
+      plotType: 'mitochondrialFractionLogHistogram',
+      plot: (config, plotData) => (<MitochondrialFractionLogHistogram experimentId={experimentId} config={config} plotData={plotData} />),
+    },
+  };
+
+  const config = useSelector((state) => state.componentConfig[plots[selectedPlot].plotUuid]?.config);
+  const plotData = useSelector((state) => state.componentConfig[plots[selectedPlot].plotUuid]?.plotData);
+
+  useEffect(() => {
+    const { plotUuid, plotType } = plots[selectedPlot];
+
+    if (!config) {
+      dispatch(loadPlotConfig(experimentId, plotUuid, plotType));
     }
-    if (config.plotToDraw) {
-      return {
-        width: config.width,
-        height: config.height,
-        autosize: { type: 'fit', resize: true },
-        padding: 5,
+  }, [selectedPlot]);
 
-        data: [
-          {
-            name: 'plotData',
-          },
-          {
-            name: 'binned',
-            source: 'plotData',
-            transform: [
-              {
-                type: 'bin',
-                field: 'fracMito',
-                extent: [0, 1],
-                step: config.binStep,
-                nice: false,
-              },
-              {
-                type: 'aggregate',
-                key: 'bin0',
-                groupby: ['bin0', 'bin1'],
-                fields: ['bin0'],
-                ops: ['count'],
-                as: ['count'],
-              },
-              {
-                type: 'formula',
-                as: 'count',
-                expr: 'datum.count/10000',
-              },
-              {
-                type: 'formula',
-                as: 'status',
-                expr: colorExpression,
-              },
-            ],
-          },
-        ],
+  useEffect(() => {
+    if (config && plotData) {
+      setPlot(plots[selectedPlot].plot(config, plotData));
+    }
+  }, [config, plotData]);
 
-        scales: [
-          {
-            name: 'xscale',
-            type: 'linear',
-            range: 'width',
-            domain: [0, 1],
-          },
-          {
-            name: 'yscale',
-            type: 'linear',
-            range: 'height',
-            round: true,
-            domain: { data: 'binned', field: 'count' },
-            zero: true,
-            nice: true,
-          },
-          {
-            name: 'color',
-            type: 'ordinal',
-            range:
-              [
-                'green', 'blue',
-              ],
-            domain: {
-              data: 'binned',
-              field: 'status',
-              sort: false,
-            },
-          },
-        ],
+  const plotStylingConfig = [
+    {
+      panelTitle: 'Legend',
+      controls: ['legend'],
+    },
+    {
+      panelTitle: 'Plot Dimensions',
+      controls: ['dimensions'],
+    },
+    {
+      panelTitle: 'Axes',
+      controls: ['axes'],
+    },
+    {
+      panelTitle: 'Title',
+      controls: ['title'],
+    },
+    {
+      panelTitle: 'Font',
+      controls: ['font'],
+    },
+  ];
 
-        axes: [
-          {
-            orient: 'bottom',
-            scale: 'xscale',
-            grid: true,
-            zindex: 1,
-            title: { value: config.xAxisText },
-            titleFont: { value: config.masterFont },
-            labelFont: { value: config.masterFont },
-            titleFontSize: { value: config.axisTitlesize },
-            labelFontSize: { value: config.axisTicks },
-            offset: { value: config.axisOffset },
-            gridOpacity: { value: (config.transGrid / 20) },
-          },
-          {
-            orient: 'left',
-            scale: 'yscale',
-            tickCount: 5,
-            grid: true,
-            zindex: 1,
-            title: { value: config.yAxisText },
-            titleFont: { value: config.masterFont },
-            labelFont: { value: config.masterFont },
-            titleFontSize: { value: config.axisTitlesize },
-            labelFontSize: { value: config.axisTicks },
-            offset: { value: config.axisOffset },
-            gridOpacity: { value: (config.transGrid / 20) },
-          },
-        ],
-
-        marks: [
-          {
-            type: 'rect',
-            from: { data: 'binned' },
-            encode: {
-              update: {
-                x: { scale: 'xscale', field: 'bin0' },
-                x2: {
-                  scale: 'xscale',
-                  field: 'bin1',
-                },
-                y: { scale: 'yscale', field: 'count' },
-                y2: { scale: 'yscale', value: 0 },
-                fill: {
-                  scale: 'color',
-                  field: 'status',
-                },
-              },
-            },
-          },
-          {
-            type: 'rect',
-            from: { data: 'plotData' },
-            encode: {
-              enter: {
-                x: { scale: 'xscale', field: 'fracMito' },
-                width: { value: 1 },
-                y: { value: 25, offset: { signal: 'height' } },
-                height: { value: 5 },
-                fillOpacity: { value: 0.4 },
-                fill: {
-                  scale: 'color',
-                  field: 'status',
-                },
-              },
-            },
-          },
-          {
-            type: 'rule',
-            encode: {
-              update: {
-                x: { scale: 'xscale', value: config.maxFraction },
-                y: { value: 0 },
-                y2: { field: { group: 'height' } },
-                strokeWidth: { value: 2 },
-                strokeDash: { value: [8, 4] },
-                stroke: { value: 'red' },
-              },
-            },
-          },
-        ],
-        legends: legend,
-        title:
-        {
-          text: { value: config.titleText },
-          anchor: { value: config.titleAnchor },
-          font: { value: config.masterFont },
-          dx: 10,
-          fontSize: { value: config.titleSize },
-        },
-      };
+  const renderPlot = () => {
+    // Spinner for main window
+    if (!config || !plotData) {
+      return (
+        <center>
+          <Loader experimentId={experimentId} />
+        </center>
+      );
     }
 
-    return {
-      width: config.width,
-      height: config.height,
-      padding: 5,
-      autosize: { type: 'fit', resize: true },
-      data: [
-        {
-          name: 'plotData',
-        },
-        {
-          name: 'binned',
-          source: 'plotData',
-          transform: [
-            {
-              type: 'bin',
-              field: 'cellSize',
-              extent: [0, 6],
-              step: config.binStep,
-              nice: false,
-            },
-            {
-              type: 'aggregate',
-              key: 'bin0',
-              groupby: ['bin0', 'bin1'],
-              fields: ['fracMito'],
-              ops: ['average'],
-              as: ['averageFracMito'],
-            },
-            {
-              type: 'formula',
-              as: 'status',
-              expr: colorExpression2,
-            },
-          ],
-        },
-      ],
+    if (plot) {
+      return plot;
+    }
+  };
 
-      scales: [
-        {
-          name: 'xscale',
-          type: 'linear',
-          range: 'width',
-          domain: [1, 4.5],
-          domainMin: 1,
-        },
-        {
-          name: 'yscale',
-          type: 'linear',
-          range: 'height',
-          round: true,
-          domain: { data: 'binned', field: 'averageFracMito' },
-          zero: true,
-          nice: true,
-        },
-        {
-          name: 'color',
-          type: 'ordinal',
-          range:
-            [
-              'blue', 'green', 'grey',
-            ],
-          domain: {
-            data: 'binned',
-            field: 'status',
-            sort: true,
-          },
-        },
-      ],
+  return (
+    <>
+      <Row>
+        <Col span={15}>
+          {renderPlot()}
+        </Col>
 
-      axes: [
-        {
-          orient: 'bottom',
-          scale: 'xscale',
-          zindex: 1,
-          grid: true,
-          title: { value: config.xAxisText2 },
-          titleFont: { value: config.masterFont },
-          labelFont: { value: config.masterFont },
-          titleFontSize: { value: config.axisTitlesize },
-          labelFontSize: { value: config.axisTicks },
-          offset: { value: config.axisOffset },
-          gridOpacity: { value: (config.transGrid / 20) },
-
-        },
-        {
-          orient: 'left',
-          scale: 'yscale',
-          tickCount: 5,
-          grid: true,
-          zindex: 1,
-          title: { value: config.yAxisText2 },
-          titleFont: { value: config.masterFont },
-          labelFont: { value: config.masterFont },
-          titleFontSize: { value: config.axisTitlesize },
-          labelFontSize: { value: config.axisTicks },
-          offset: { value: config.axisOffset },
-          gridOpacity: { value: (config.transGrid / 20) },
-        },
-      ],
-
-      marks: [
-        {
-          type: 'rect',
-          from: { data: 'binned' },
-          encode: {
-            update: {
-              x: { scale: 'xscale', field: 'bin0' },
-              x2: {
-                scale: 'xscale',
-                field: 'bin1',
-              },
-              y: { scale: 'yscale', field: 'averageFracMito' },
-              y2: { scale: 'yscale', value: 0 },
-              fill: {
-                scale: 'color',
-                field: 'status',
-              },
-            },
-          },
-        },
-        {
-          type: 'rule',
-          encode: {
-            update: {
-              x: { scale: 'xscale', value: config.maxFraction2 },
-              y: { value: 0 },
-              y2: { field: { group: 'height' } },
-              strokeWidth: { value: 2 },
-              strokeDash: { value: [8, 4] },
-              stroke: { value: 'red' },
-            },
-          },
-        },
-      ],
-      legends: legend,
-
-      title:
-      {
-        text: { value: config.titleText },
-        anchor: { value: config.titleAnchor },
-        font: { value: config.masterFont },
-        dx: 10,
-        fontSize: { value: config.titleSize },
-      },
-    };
-  }
-
-  render() {
-    const data = { plotData: this.generateData() };
-    const { config } = this.state;
-    // eslint-disable-next-line react/prop-types
-    const {
-      experimentId, sampleId, filtering, sampleIds,
-    } = this.props;
-
-    const changePlot = (val) => {
-      this.updatePlotWithChanges({ plotToDraw: val });
-      if (val) {
-        this.updatePlotWithChanges({
-          xDefaultTitle: config.xAxisText,
-          yDefaultTitle: config.yAxisText,
-          placeholder: 0.1,
-          sliderMax: 1,
-          sliderMin: 0,
-        });
-      } else {
-        this.updatePlotWithChanges({
-          xDefaultTitle: config.xAxisText2,
-          yDefaultTitle: config.yAxisText2,
-          placeholder: 3.5,
-          sliderMax: 4.5,
-          sliderMin: 1,
-        });
-      }
-    };
-
-    return (
-      <>
-        <Row>
-
-          <Col span={15}>
-            <Vega data={data} spec={this.generateSpec()} renderer='canvas' />
-          </Col>
-
-          <Col span={3}>
-            <Space direction='vertical'>
-              <Tooltip title='A high fraction of mitochondrial reads is an indicator of cell death. The usual range for this cut-off is 0.1-0.5.'>
-                <Button icon={<InfoCircleOutlined />} />
-              </Tooltip>
-              <img
-                alt=''
-                src={plot1Pic}
+        <Col span={3}>
+          <Space direction='vertical'>
+            <Tooltip title='A high fraction of mitochondrial reads is an indicator of cell death. The usual range for this cut-off is 0.1-0.5.'>
+              <Button icon={<InfoCircleOutlined />} />
+            </Tooltip>
+            {Object.entries(plots).map(([key, option]) => (
+              <button
+                type='button'
+                key={key}
+                onClick={() => setSelectedPlot(key)}
                 style={{
-                  height: '100px',
-                  width: '100px',
-                  align: 'center',
-                  padding: '8px',
-                  border: '1px solid #000',
+                  padding: 0, margin: 0, border: 0, backgroundColor: 'transparent',
                 }}
-                onClick={() => changePlot(true)}
-              />
-              <img
-                alt=''
-                src={plot2Pic}
-                style={{
-                  height: '100px',
-                  width: '100px',
-                  align: 'center',
-                  padding: '8px',
-                  border: '1px solid #000',
-                }}
-                onClick={() => changePlot(false)}
-              />
-            </Space>
-          </Col>
+              >
+                <img
+                  alt={option.title}
+                  src={option.imgSrc}
+                  style={{
+                    height: '100px',
+                    width: '100px',
+                    align: 'center',
+                    padding: '8px',
+                    border: '1px solid #000',
+                  }}
+                />
+              </button>
 
-          <Col span={6}>
-            <Space direction='vertical' style={{ width: '100%' }} />
-            <Collapse defaultActiveKey={['filtering-settings']}>
-              <Panel header='Filtering settings' collapsible={!filtering ? 'disabled' : 'header'} key='filtering-settings'>
-                <CalculationConfig experimentId={experimentId} sampleId={sampleId} sampleIds={sampleIds} />
-              </Panel>
+            ))}
+          </Space>
+        </Col>
 
-              {/* Temporary placeholder, replace with <PlotStyling> when working on this component */}
-              <OldPlotStyling
-                config={config}
-                onUpdate={this.updatePlotWithChanges}
-                updatePlotWithChanges={this.updatePlotWithChanges}
-                legendMenu
-              />
-            </Collapse>
-          </Col>
-        </Row>
-      </>
-    );
-  }
-}
+        <Col span={6}>
+          <Collapse defaultActiveKey={['settings']}>
+            <Panel header='Filtering Settings' key='settings'>
+              <CalculationConfig experimentId={experimentId} />
+            </Panel>
+            <Panel header='Plot styling' key='styling'>
+              <div style={{ height: 8 }} />
+              <PlotStyling formConfig={plotStylingConfig} config={config} onUpdate={updatePlotWithChanges} />
+            </Panel>
+          </Collapse>
+        </Col>
+      </Row>
+    </>
+  );
+};
 
 MitochondrialContent.propTypes = {
   experimentId: PropTypes.string.isRequired,
-  sampleId: PropTypes.string.isRequired,
-  filtering: PropTypes.bool,
-  sampleIds: PropTypes.array.isRequired,
-};
-
-MitochondrialContent.defaultProps = {
-  filtering: true,
 };
 
 export default MitochondrialContent;
